requirements.txt
pip
connexion
connexion[flask]
connexion[uvicorn]
connexion[swagger-ui]
Flask-SQLAlchemy
Flask-Cors
PyMySQL
alembic
bcrypt
pytz
feedgen
tabulate
